requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
dnspython==2.7.0
eventlet==0.33.3
Flask==2.3.3
Flask-SocketIO==5.3.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.63.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-engineio==4.11.2
python-socketio==5.10.0
referencing==0.36.2
rpds-py==0.23.1
rti.connext==7.3.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==2.3.7
wsproto==1.2.0
